+++ epoxy-processor/src/main/java/com/airbnb/epoxy/processor/PackageModelViewSettings.kt
@@ -16,6 +16,10 @@ class PackageModelViewSettings(
     val layoutName: String = annotation.defaultLayoutPattern
     val includeAlternateLayouts: Boolean = annotation.useLayoutOverloads
     val generatedModelSuffix: String = annotation.generatedModelSuffix
+    val disableGenerateBuilderOverloads: Boolean? =
+        annotation.disableGenerateBuilderOverloads.toBoolean()
+    val disableGenerateGetters: Boolean? = annotation.disableGenerateGetters.toBoolean()
+    val disableGenerateReset: Boolean? = annotation.disableGenerateReset.toBoolean()
 
     val defaultBaseModel: TypeMirror? by lazy {
 
@@ -40,5 +44,13 @@ class PackageModelViewSettings(
         val viewName = Utils.toSnakeCase(viewElement.simpleName.toString())
         val resourceName = layoutName.replace("%s", viewName)
         return ResourceValue(rClass, resourceName, 0)
+    }
+
+    private fun PackageModelViewConfig.Option.toBoolean(): Boolean? {
+        return when (this) {
+            PackageModelViewConfig.Option.Default -> null
+            PackageModelViewConfig.Option.Enabled -> true
+            PackageModelViewConfig.Option.Disabled -> false
+        }
     }
 }