+++ epoxy-processortest/src/test/resources/TestCallbackPropViewModel_.java
@@ -8,13 +8,10 @@ import java.lang.Number;
 import java.lang.Object;
 import java.lang.Override;
 import java.lang.String;
-import java.util.BitSet;
 
 /**
  * Generated file. Do not modify! */
 public class TestCallbackPropViewModel_ extends EpoxyModel<TestCallbackPropView> implements GeneratedModel<TestCallbackPropView>, TestCallbackPropViewModelBuilder {
-  private final BitSet assignedAttributes_epoxyGeneratedModel = new BitSet(1);
-
   private OnModelBoundListener<TestCallbackPropViewModel_, TestCallbackPropView> onModelBoundListener_epoxyGeneratedModel;
 
   private OnModelUnboundListener<TestCallbackPropViewModel_, TestCallbackPropView> onModelUnboundListener_epoxyGeneratedModel;
@@ -23,8 +20,6 @@ public class TestCallbackPropViewModel_ extends EpoxyModel<TestCallbackPropView>
 
   private OnModelVisibilityChangedListener<TestCallbackPropViewModel_, TestCallbackPropView> onModelVisibilityChangedListener_epoxyGeneratedModel;
 
-  /**
-   * Bitset index: 0 */
   @Nullable
   private View.OnClickListener listener_OnClickListener = (View.OnClickListener) null;
 
@@ -152,7 +147,6 @@ public class TestCallbackPropViewModel_ extends EpoxyModel<TestCallbackPropView>
    * Set a click listener that will provide the parent view, model, and adapter position of the clicked view. This will clear the normal View.OnClickListener if one has been set */
   public TestCallbackPropViewModel_ listener(
       @Nullable final OnModelClickListener<TestCallbackPropViewModel_, TestCallbackPropView> listener) {
-    assignedAttributes_epoxyGeneratedModel.set(0);
     onMutation();
     if (listener == null) {
       this.listener_OnClickListener = null;
@@ -169,7 +163,6 @@ public class TestCallbackPropViewModel_ extends EpoxyModel<TestCallbackPropView>
    * @see TestCallbackPropView#setListener(View.OnClickListener)
    */
   public TestCallbackPropViewModel_ listener(@Nullable View.OnClickListener listener) {
-    assignedAttributes_epoxyGeneratedModel.set(0);
     onMutation();
     this.listener_OnClickListener = listener;
     return this;
@@ -260,7 +253,6 @@ public class TestCallbackPropViewModel_ extends EpoxyModel<TestCallbackPropView>
     onModelUnboundListener_epoxyGeneratedModel = null;
     onModelVisibilityStateChangedListener_epoxyGeneratedModel = null;
     onModelVisibilityChangedListener_epoxyGeneratedModel = null;
-    assignedAttributes_epoxyGeneratedModel.clear();
     this.listener_OnClickListener = (View.OnClickListener) null;
     super.reset();
     return this;